[AIP-21] Move operators.docker_operator providers.docker.operators.docker
