+++ airflow/providers/docker/operators/docker_swarm.py
@@ -20,7 +20,7 @@
 from docker import types
 
 from airflow.exceptions import AirflowException
-from airflow.operators.docker_operator import DockerOperator
+from airflow.providers.docker.operators.docker import DockerOperator
 from airflow.utils.decorators import apply_defaults
 from airflow.utils.strings import get_random_string
