+++ tests/providers/docker/operators/test_docker_swarm.py
@@ -28,7 +28,7 @@ from airflow.providers.docker.operators.docker_swarm import DockerSwarmOperator
 
 class TestDockerSwarmOperator(unittest.TestCase):
 
-    @mock.patch('airflow.operators.docker_operator.APIClient')
+    @mock.patch('airflow.providers.docker.operators.docker.APIClient')
     @mock.patch('airflow.providers.docker.operators.docker_swarm.types')
     def test_execute(self, types_mock, client_class_mock):
 
@@ -81,7 +81,7 @@ class TestDockerSwarmOperator(unittest.TestCase):
         self.assertEqual(client_mock.tasks.call_count, 3)
         client_mock.remove_service.assert_called_once_with('some_id')
 
-    @mock.patch('airflow.operators.docker_operator.APIClient')
+    @mock.patch('airflow.providers.docker.operators.docker.APIClient')
     @mock.patch('airflow.providers.docker.operators.docker_swarm.types')
     def test_no_auto_remove(self, types_mock, client_class_mock):
 
@@ -107,7 +107,7 @@ class TestDockerSwarmOperator(unittest.TestCase):
             'Docker service being removed even when `auto_remove` set to `False`'
         )
 
-    @mock.patch('airflow.operators.docker_operator.APIClient')
+    @mock.patch('airflow.providers.docker.operators.docker.APIClient')
     @mock.patch('airflow.providers.docker.operators.docker_swarm.types')
     def test_failed_service_raises_error(self, types_mock, client_class_mock):
